requirements.txt
fastapi[standard]
uvicorn[standard]
ultralytics
opencv-python
pillow
python-dotenv
